add read sequence for 6 button pce controllers

+++ headers/addons/tg16_input.h
@@ -65,7 +65,7 @@ private:
 
     uint16_t map(uint16_t x, uint16_t in_min, uint16_t in_max, uint16_t out_min, uint16_t out_max);
     uint8_t readController();
-    void updateButtons(uint8_t data);
+    void updateButtons(uint16_t data);
 };
 
 #endif  // _TG16padAddon_H
